refactor(BackToTopButton): move to own component folder
